+++ eras/allegra/impl/testlib/Test/Cardano/Ledger/Allegra/Era.hs
@@ -1,4 +1,6 @@
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -12,11 +14,17 @@ import Cardano.Ledger.Allegra.Core
 import Cardano.Ledger.Allegra.Scripts
 import Cardano.Ledger.MemoBytes (EqRaw)
 import Cardano.Ledger.Plutus (emptyCostModels)
+import Paths_cardano_ledger_allegra (getDataFileName)
 import Test.Cardano.Ledger.Allegra.Arbitrary ()
 import Test.Cardano.Ledger.Allegra.Binary.Annotator ()
+import Test.Cardano.Ledger.Allegra.Examples (exampleAllegraTx)
 import Test.Cardano.Ledger.Allegra.TreeDiff ()
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Shelley.Era
+import Test.Cardano.Ledger.Shelley.Examples (
+  exampleShelleyPParams,
+  exampleShelleyPParamsUpdate,
+ )
 
 class
   ( ShelleyEraTest era
@@ -31,11 +39,33 @@ class
   AllegraEraTest era
 
 instance EraTest AllegraEra where
+  type
+    EraRulesWithFailures AllegraEra =
+      '[ "BBODY"
+       , "DELEG"
+       , "DELEGS"
+       , "DELPL"
+       , "LEDGER"
+       , "LEDGERS"
+       , "POOL"
+       , "PPUP"
+       , "UTXO"
+       , "UTXOW"
+       ]
+
   zeroCostModels = emptyCostModels
 
   mkTestAccountState = mkShelleyTestAccountState
 
   accountsFromAccountsMap = shelleyAccountsFromAccountsMap
+
+  mkEraFullPath = getDataFileName
+
+  exampleTx = exampleAllegraTx
+
+  examplePParams = exampleShelleyPParams
+
+  examplePParamsUpdate = exampleShelleyPParamsUpdate
 
 instance ShelleyEraTest AllegraEra