+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Gov.hs
@@ -5,42 +5,27 @@
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Gov () where
 
-import Cardano.Ledger.BaseTypes
-import Cardano.Ledger.Conway.Core
-import Cardano.Ledger.Conway.Governance
-import Cardano.Ledger.Conway.Rules
+import Cardano.Ledger.Conway (ConwayEra)
+import qualified Cardano.Ledger.Conway.Rules as Conway
 import Cardano.Ledger.State
-import Data.Functor.Identity (Identity)
 import qualified Data.Map.Strict as Map
 import Lens.Micro ((^.))
-import qualified MAlonzo.Code.Ledger.Foreign.API as Agda
+import qualified MAlonzo.Code.Ledger.Conway.Foreign.API as Agda
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Base
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Base ()
+import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Cert ()
 
-instance
-  ( SpecTranslate (PParamsHKD Identity era)
-  , SpecContext (PParamsHKD Identity era) ~ ()
-  , EraPParams era
-  , SpecRep (PParamsHKD Identity era) ~ Agda.PParams
-  , SpecTranslate (CertState era)
-  , SpecContext (CertState era) ~ ()
-  , SpecRep (CertState era) ~ Agda.CertState
-  , EraCertState era
-  ) =>
-  SpecTranslate (GovEnv era)
-  where
-  type SpecRep (GovEnv era) = Agda.GovEnv
-  type SpecContext (GovEnv era) = EnactState era
+instance SpecTranslate ConwayEra (Conway.GovEnv ConwayEra) where
+  type SpecRep ConwayEra (Conway.GovEnv ConwayEra) = Agda.GovEnv
+  type SpecContext ConwayEra (Conway.GovEnv ConwayEra) = Conway.EnactState ConwayEra
 
-  toSpecRep GovEnv {..} = do
+  toSpecRep Conway.GovEnv {..} = do
     enactState <- askSpecTransM
     let rewardAccounts = Map.keysSet $ geCertState ^. certDStateL . accountsL . accountsMapL
     withCtxSpecTransM () $
@@ -53,17 +38,10 @@ instance
         <*> toSpecRep geCertState
         <*> toSpecRep rewardAccounts
 
-instance
-  ( EraPParams era
-  , SpecTranslate (PParamsHKD StrictMaybe era)
-  , SpecContext (PParamsHKD StrictMaybe era) ~ ()
-  , SpecRep (PParamsHKD StrictMaybe era) ~ Agda.PParamsUpdate
-  ) =>
-  SpecTranslate (GovSignal era)
-  where
-  type SpecRep (GovSignal era) = [Either Agda.GovVote Agda.GovProposal]
+instance SpecTranslate ConwayEra (Conway.GovSignal ConwayEra) where
+  type SpecRep ConwayEra (Conway.GovSignal ConwayEra) = [Either Agda.GovVote Agda.GovProposal]
 
-  toSpecRep GovSignal {gsVotingProcedures, gsProposalProcedures} = do
+  toSpecRep Conway.GovSignal {gsVotingProcedures, gsProposalProcedures} = do
     votingProcedures <- toSpecRep gsVotingProcedures
     proposalProcedures <- toSpecRep gsProposalProcedures
     pure $